make it works with the typedefs of libdecompress.so

+++ src/defs.h
@@ -3,7 +3,8 @@
 
 typedef unsigned char       byte   ;
 typedef unsigned short      uint16 ;
-typedef unsigned long       uint32 ;
+//typedef unsigned long       uint32 ;
+typedef unsigned int       uint32 ;
 typedef unsigned long long  uint64 ;
 
 typedef unsigned long       dostime_t;
